backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai>=1.30.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pillow==10.3.0
sentence-transformers==2.6.1
torch>=2.0.0
numpy>=1.26.0
python-multipart==0.0.9
requests==2.31.0
beautifulsoup4==4.12.3
schedule==1.2.1
pydantic==2.6.4
pydantic-settings==2.2.1
playwright==1.40.0
